Remove setFilteredItems and getFilteredItems
Breaking change!

This API is un-tested and not used in any demo. The use-cases are
unclear. In the webcomponent, filteredItems was used to allow custom
filtering with items on filter-changed event. However, adding a
listener for this event is not public API. Also, the javadoc of the
getter is wrong: the property doesn't contain currently filtered items,
unless the developer has explicitly set those.

If the purpose of the setter was to allow custom filtering (even though
you can't listen for the event), it's redundant now that we have
ItemFilter.