Use 'value' as item identifier to fix BFP
Fixes https://github.com/vaadin/flow/issues/4862

This is a workaround for a certain situation when client-side
value-setter is not called when the combo boxes are wrapped in a certain
PolymerTemplate structure.

Now the item.value is used as itemIdPath, which means that when setting
the selectedItem, the web component finds the value identifier
immediately from that item.

Previously it worked like so that later after setting selectedItem, the
value-property is set which fixed everything before the itemIdPath was
set. But this value-property is not going through the Polymer
property-setting in this certain scenario.

This causes some extra overhead: we need the old item.key also because
the datagenerators and renderers depend on it. So there will be
duplicate property in each data item: key and value. We need to consider if this is
worth it to fix the issue.