Add maven profile to use bower in integration tests
This allows installing any GitHub branch of the web component, to test how
the changes in the web component affect the Flow counterpart, by
defining the property -DuseBower. The branch can be specified in the
added bower.json (default is master).
Running 'mvn clean' will remove the bower components.

The configuration is based on vaadin-crud-flow.