Update value-property on setValue
Fix #216

The problem was that the server-side value-property did not match the
key of the set item. When the value-prop was later set with js as a
workaround for a Flow bug (see the lines below the added line in
ComboBox.java), that value did not match the selectedItem, which caused
the value to become null.

(Note: server-side key is used as the client-side value-property)