Prevent default on custom-value-set, improve javadocs
If preventDefault is not called, the web component will automatically
set the custom value as the value-property of combo box. This breaks the
Flow component, which uses the key of the currently selected Java bean
as the value-property.

Docs are improved to explain that it's the developer's responsibility to
handle the custom-value-set events, as we can't automatically convert
the custom string to an arbitrary bean type.

Integration tests are added, as previously none existed for this
feature.

Fix #236