Use Element context on JS call to init connector vaadin/vaadin-combo-box#272
This avoids it to be executed when the element doesn't exist in
client-side. Also, removed unnecessary call to connector.reset, which
will be anyway handled through refreshAllData().

Fixes: 

This somehow causes "$connector.enableClientValidation" to be called
before connector.init, so it had to be postponed with beforeClientResponse.

Also fixed false-positive disabled IT. It now errored because combo-box
doesn't have filteredItems at all now before the server has responded with
some data.

Flow-component: vaadin-combo-box